Starts creating check coordinates methods, psuedocodes further plans
