+++ battleship.rb
@@ -1,57 +1,113 @@
-puts "Welcome to Battleship, Player One."
+require 'pry'
 
-puts "Would you like to (p)lay, read the (i)nstructions, or (q)uit?"
-answer = gets.chomp
+class Welcome
 
-# this needs to be looped, currently this is not how that works though
-
-
-  if answer == "p" || answer == "play"
-    # computer places ship based on spec parameters
-    # Ships cannot wrap around the board
-    # Ships cannot overlap
-    # Ships can be laid either horizontally or vertically
-    # Coordinates must correspond to the first and last units
-    # of the ship. (IE: You can’t place a two unit ship at “A1 A3”)
-    # figure out how to automatically have the computer ships put in place
-    puts "Computer player: 'Hello, honored adversary, my armada is in place.
-    Place your two ships.
-    Your first is two units long, and the second is three units long.
-    The grid has A1 at the top left, and D4 at the bottom right.'
-    Enter the squares for your two unit ship:"
+  def welcome_page
+    # binding.pry
+    puts "Welcome to Battleship, Player One."
+    puts "Would you like to (p)lay, read the (i)nstructions, or (q)uit?"
     answer = gets.chomp
-    # make sure this is only two units and that it must be enterd before moving forward
-    # if loop for answer parameters and guidelines
-    puts "Now enter the squares for your three unit ship:"
-    answer = gets.chomp
-    # if loop for answer parameters and guidelines
-    # make sure this is three units
-    # we need to validate the entries according to the parameters
-    # THIS IS WHERE WE START main game flow phase
-  elsif
-    answer == "i" || answer == "instructions"
-    puts "1. Choose 'p' to play game.
-    2. You and computer player set ships.
-    3. Take turns firing missiles at each others grids the idea is to destroy your enemies ships before they destroy yours.
-    4. If you hit a ship, it will be indicated with an 'H', a miss will be indicated with an 'M'.
-    5. Whoever sinks both of their opponent's ships wins!!"
-
-    puts "Would you like to (p)lay or (q)uit?"
-    answer = gets.chomp
-    # also needs to loop, currently goes into the ether
-    # loop in a loop
-  elsif
-    answer == "q" || answer == "quit"
-    puts "Boring!"
-  else
-    p "Invalid entry, please try again."
-    # needs to loop
+    if answer == "p"
+      play
+    elsif
+      answer == "i"
+      instructions
+    elsif
+      answer == "q"
+      quit
+    else
+      invalid_input
+    end
   end
 
 
-# ^^^^will need to be some sort of loop so that IF it is an
-# invalid entry, it will restart at the top. REPL LOL
-# also...how do we make this "go somewhere" when they select "play"
-# it seems like it can all be in here, confused about classes or
-# if this is just something that can be run like this
-# have not yet delved into how it actually works with everything IN CODE
+  def game_setup
+    size = 1
+    player_board = Board.new(size)
+    computer_board = Board.new(size)
+    @game = Game.new(player_board, computer_board)
+  end
+
+  def place_first_ship(coordinates)
+    # call parameter method
+    if @game.player_board.validation_station(coordinates)
+    #place_the_first_ship
+    else
+      invalid_input
+      place_coordinates
+    end
+  end
+
+  def place_second_ship(coordinates)
+    # call parameter method
+  end
+
+
+  def place_coordinates
+    puts "Ships cannot wrap around the board."
+    puts "Ships cannot overlap"
+    puts "Ships can be laid either horizontally or vertically"
+    puts "Coordinates must correspond to the first and last units of the ship. \n (IE: You can’t place a two unit ship at “A1 A3”)"
+    puts "Enter the coordinates for your two unit ship:"
+    game_setup
+    placement_1 = gets.chomp
+    # validate first coordinate
+    # place 2-unit_ship
+
+    puts "Now enter the coordinates for your three unit ship:"
+    placement_2 = gets.chomp
+    #validate second coordinates
+    # get the board to tell us if these are valid
+    #if invalid say invalid and call this method again LoopLife
+  end
+
+  def play
+    puts "Computer player: 'Hello, honored adversary, my armada is in place.
+    Place your two ships.
+    Your first is two units long, and the second is three units long.
+    The grid has A1 at the top left, and D4 at the bottom right.' "
+    place_coordinates
+  end
+
+  def invalid_input
+    puts "Invalid entry."
+  end
+
+  def instructions
+    puts "1. Choose 'p' to play game."
+    puts "2. You and computer player set ships."
+    puts "3. Take turns firing missiles at each others grids the idea is \n to destroy your enemies ships before they destroy yours."
+    puts "4. If you hit a ship, it will be indicated with an 'H', a miss will be indicated with an 'M'."
+    puts "5. Whoever sinks both of their opponent's ships wins!!"
+    # binding.pry
+    puts "Would you like to (p)lay or (q)uit?"
+    choice = gets.chomp
+    if choice == "p"
+      # binding.pry
+      play
+    elsif choice == "q"
+      # binding.pry
+      quit
+    else
+      # binding.pry
+      invalid_input
+      instructions
+    end
+  end
+
+  def quit
+    puts "Are you sure you aren't ready for battle? (q)uit or (p)lay."
+    reply = gets.chomp
+    if reply == "q"
+      puts "My armada will await you for another time."
+    elsif reply == "p"
+      play
+    else
+      invalid_input
+    end
+  end
+
+end
+
+welcome = Welcome.new
+welcome.welcome_page
