+++ lib/board.rb
@@ -1,14 +1,52 @@
 require './lib/space'
 
 class Board
-  attr_accessor :grid
 
-# work in progress
-  def initialize(square = Space.new)
-    @grid = []
-    @grid << square
-    # hash....? who builds the board
-    # binding.pry
+  def initialize(board_size)
+    @board_size = board_size
   end
+
+  def validation_station(coordinates, ship_size)
+    # whether or not the thing is on the board
+    # return true/false
+    #
+    # know about ship size
+  end
+
+  def check_both(coordinates)
+    #coordinates.all? do |coordinate|
+      #are in list of valid coordinates
+      #on_board?(coordinate)
+      #
+  end
+
+  def on_board?(coordinate)
+    #will check if on list of valid coordinates
+    #if coordinate is in array of valid coordinates, array.include?
+    
+  end
+
+# method that checks if the first and last coordinate are equal to ship's size
+# coordinates must correspond to the first and last units of the ship
+# method that checks if ship is not wrapping around board
+
+
+
+
+
+#ship cannot wrap around the board
+#ship cannot be on top of another ship
+# ship cannot be placed diagonally
+
+
+
+
+
+
+
+
+
+
+
 
 end
