+++ lib/game.rb
@@ -1,2 +1,13 @@
 class Game
+  attr_reader :player_board
+
+  def initialize(player_board, computer_board)
+    @player_board = player_board
+    @computer_board = computer_board
+  end
+
+
+
+
+
 end
